pyproject.toml
[tool.poetry]
name = "whatsapi"
version = "0.1.0"
description = ""
authors = ["Carlos Moreno <dev@seekingaura.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pywhatkit = "^5.4"
django = "^5.0.4"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
dj-database-url = "^2.1.0"
celery = {extras = ["redis"], version = "^5.3.6"}
django-celery-results = "^2.5.1"
django-celery-beat = "^2.6.0"

[tool.poetry.group.server]
optional = true

[tool.poetry.group.server.dependencies]
uwsgi = "^2.0.24"

[tool.black]
line-length = 119

[tool.isort]
force_grid_wrap = 2
multi_line_output = 3
include_trailing_comma = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
